requirements.txt
# Core FastAPI dependencies
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Environment configuration
python-dotenv==1.0.0

# HTTP client for testing
requests==2.31.0

# Redis for conversation state management (includes async support)
redis==5.0.1

# OpenAI for STT, LLM, TTS
openai==1.12.0

# Asterisk ARI (REST Interface) client
ari-py==0.1.3
websockets==12.0

# Audio processing
pydub==0.25.1

# Authentication - Supabase JWT
PyJWT==2.8.0
cryptography==42.0.1

# Development tools
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
